+++ contribution/views.py
@@ -4,7 +4,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.http import HttpResponse
-from contribution.models import Contributor
+from .models import Contributor
 from account.serializers import PersonSerializer
 from .serializers import ContributorSerializer
 from rest_framework import viewsets
@@ -14,93 +14,58 @@ from rest_framework.permissions import IsAuthenticated
 # Create your views here.
 
 
-class IsAuthor(permissions.BasePermission):
-    message = "You cannot remove the author of this project."
+class IsContributorOrAuthor(permissions.BasePermission):
+    message = "You must be the project author or contributor."
 
     def has_object_permission(self, request, view, obj):
-        return not obj.permission == "author"
-
-
-class ProjectContributors(APIView):
-    def get_project(self, id):
-        try:
-            return Project.objects.get(id=id)
-
-        except Project.DoesNotExist:
-            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
-
-    def get(self, request, id):
-        project = self.get_project(id)
-
-        if Contributor.objects.filter(user=request.user, project=project):
-            contributors = Contributor.objects.filter(project=project)
-
-            persons = [contributor.user for contributor in contributors]
-            serializer = PersonSerializer(persons, many=True)
-            return Response(serializer.data)
+        if Contributor.objects.filter(project=obj, user=request.user):
+            return True
         else:
-            return Response(status=status.HTTP_401_UNAUTHORIZED)
-
-    def post(self, request, id):
-        project = self.get_project(id)
-
-        if Contributor.objects.filter(user=request.user, project=project):
-            data = request.data.copy()
-            data["user"] = Person.objects.get(username=data["user"]).id
-            serializer = ContributorSerializer(data=data)
-            if serializer.is_valid():
-                serializer.save(permission="contributor", role="Contributor")
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-        return Response(status=status.HTTP_401_UNAUTHORIZED)
+            return False
 
 
-class DeleteContributor(APIView):
+class IsAuthor(permissions.BasePermission):
+    message = "You must be the project author."
 
-    permission_classes = [IsAuthor]
-
-    def get_project(self, id):
-        try:
-            return Project.objects.get(id=id)
-
-        except Project.DoesNotExist:
-            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
-
-    def get_user(self, id):
-        try:
-            return Person.objects.get(id=id)
-
-        except Person.DoesNotExist:
-            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
-
-    def delete(self, request, id, user_id):
-        project = self.get_project(id)
-
-        user_deleted = self.get_user(user_id)
-        if Contributor.objects.filter(user=user_deleted, project=project):
-
-            contribution_deleted = Contributor.objects.filter(
-                user=user_deleted, project=project
-            )[0]
-
-            self.check_object_permissions(self.request, contribution_deleted)
-
-            Contributor.objects.filter(user=user_deleted, project=project).delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+    def has_object_permission(self, request, view, obj):
+        if Contributor.objects.filter(project=obj, user=request.user):
+            permission = Contributor.objects.filter(project=obj, user=request.user)
+            return not permission.permission == "author"
+        else:
+            return False
 
 
 class ContributionViewSet(viewsets.ViewSet):
+    def get_permissions(self):
+        """
+        Instantiates and returns the list of permissions that this view requires.
+        """
+        if self.action == "destroy":
+            permission_classes = [IsAuthenticated, IsAuthor]
+        else:
+            permission_classes = [IsAuthenticated, IsContributorOrAuthor]
+
+        return [permission() for permission in permission_classes]
+
     def get_project(self, id):
         try:
             return Project.objects.get(id=id)
 
         except Project.DoesNotExist:
-            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
+            return None
+
+    def get_contributor(self, id):
+        try:
+            return Project.objects.get(id=id)
+
+        except Project.DoesNotExist:
+            return None
 
     def list(self, request, projects_pk):
         project = self.get_project(projects_pk)
+        if not project:
+            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
+        self.check_object_permissions(request, project)
 
         contributors = Contributor.objects.filter(project=project)
 
@@ -108,6 +73,10 @@ class ContributionViewSet(viewsets.ViewSet):
         return Response(serializer.data)
 
     def create(self, request, projects_pk):
+        project = self.get_project(projects_pk)
+        if not project:
+            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
+        self.check_object_permissions(request, project)
 
         data = request.data.copy()
         data["user"] = Person.objects.get(username=data["user"]).id
@@ -119,7 +88,17 @@ class ContributionViewSet(viewsets.ViewSet):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def destroy(self, request, projects_pk, pk=None):
-        contributor = Contributor.objects.get(id=pk)
+        project = self.get_project(projects_pk)
+        if not project:
+            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
+        self.check_object_permissions(request, project)
 
-        contributor.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        contributor = self.get_contributor(pk)
+        if not contributor:
+            return HttpResponse(status=status.HTTP_404_NOT_FOUND)
+
+        if not contributor.permission == "author":
+            contributor.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+        return Response(status=status.HTTP_400_BAD_REQUEST)
